% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{contingency_table}
\alias{contingency_table}
\title{Simmple Contingency Table Analysis}
\usage{
contingency_table(obs)
}
\arguments{
\item{obs}{this is just a contingency table. That is it.}
}
\value{

}
\description{
This function returns simple measures of indepence for I x J contingency tables. Namely, 
it gives Pearsons X^2 and the likelihood ratio test statistic, G^2. Both are asymptotically distributed 
chi-squared with I-1 x J-1 degrees of freedom. It aslo returns pearson standardized residuals for the 
contingency table.
}
\author{
Ben Brennan (brennben at umich dot edu)
}
